Bug 394890 – Segfault when running vte or gnome-terminal
2007-01-09  Ryan Lortie  <desrt@desrt.ca>

        Bug 394890 – Segfault when running vte or gnome-terminal

        * src/vtetc.c (_vte_termcap_find_string): change 'int' to 'gssize' to
        match previous fixup.  Hopefully that does it. :)


svn path=/trunk/; revision=1426

+++ src/vtetc.c
@@ -289,7 +289,7 @@ _vte_termcap_find_string (VteTermcap *termcap,
                           const char *tname,
                           const char *cap)
 {
-  int length;
+  gssize length;
 
   return _vte_termcap_find_string_length (termcap, tname, cap, &length);
 }
